requirements.txt
django

djangorestframework

apscheduler ?